Correct local.conf.sample location
This follows https://github.com/openembedded/openembedded-core/commit/214c3dc91ceb1557b62ad3a359c72cfc9b2bf4f0

+++ scripts/prepare.sh
@@ -12,6 +12,15 @@ BASEDIR=$PWD
 
 . ./oe-init-build-env $PWD/build
     
-cp $BASEDIR/meta/conf/local.conf.sample conf/local.conf
+if [ -d $BASEDIR/meta/conf/templates/default ] ; then
+	cp $BASEDIR/meta/conf/templates/default/local.conf.sample conf/local.conf
+elif [ -d $BASEDIR/meta-poky/conf/templates/default ] ; then
+	cp $BASEDIR/meta-poky/conf/templates/default/local.conf.sample conf/local.conf
+elif [ -f $BASEDIR/meta/conf/local.conf.sample ] ; then
+	cp $BASEDIR/meta/conf/local.conf.sample conf/local.conf
+else
+	echo "Unable to find a local.conf.sample"
+	exit 1
+fi
     
 echo "BBLAYERS +=\"$1/meta-meson\"" >> conf/bblayers.conf
